Get tests passing under nose2
