+++ bookie/lib/applog.py
@@ -30,7 +30,7 @@ class Log(object):
 
 class AuthLog(Log):
     """Store auth specific log items"""
-    component = "AUTH"
+    component = u"AUTH"
 
     @staticmethod
     def login(username, success, password=None):
@@ -39,9 +39,8 @@ class AuthLog(Log):
         passwd = lambda x: None if password is None else {'password': password}
 
         status = get_status(success)
-        message = "User {0} attempted to login {1}".format(username,
-                                                           success)
-
+        message = u"User {0} attempted to login {1}".format(username,
+                                                            success)
         data = {
             'user': username,
             'component': AuthLog.component,
@@ -53,7 +52,7 @@ class AuthLog(Log):
     @staticmethod
     def disabled(username):
         """Attempt to log into a disabled account"""
-        msg = "{0} is a disabled user account".format(username)
+        msg = u"{0} is a disabled user account".format(username)
 
         data = {
             'user': username,
@@ -66,9 +65,9 @@ class AuthLog(Log):
     def reactivate(username, success=True, code=None):
         """The account was marked for reactivation"""
         if success:
-            msg = "{0} was reactivated".format(username)
+            msg = u"{0} was reactivated".format(username)
         else:
-            msg = "{0} attempted to reactivate with invalid credentials"
+            msg = u"{0} attempted to reactivate with invalid credentials"
             msg = msg.format(username)
 
         LOG.debug(msg)
@@ -86,7 +85,7 @@ class AuthLog(Log):
 
 class BmarkLog(Log):
     """Bookmark specific log items"""
-    component = "BMARKS"
+    component = u"BMARKS"
 
     @staticmethod
     def export(for_user, current_user):
@@ -101,7 +100,7 @@ class BmarkLog(Log):
             current_user = "None"
 
         status = get_status(your_export)
-        message = "User {0} exported the bookmarks for {1}".format(
+        message = u"User {0} exported the bookmarks for {1}".format(
             current_user, for_user)
 
         data = {
@@ -122,7 +121,9 @@ class LogRecord(object):
 
         # we need to hash down the payload if there is one
         if 'payload' in kwargs and kwargs['payload'] is not None:
-            kwargs['payload'] = json.dumps(dict(kwargs.get('payload')))
+            kwargs['payload'] = unicode(
+                json.dumps(dict(kwargs.get('payload')))
+            )
 
         AppLogMgr.store(**kwargs)
