+++ bookie/lib/urlhash.py
@@ -5,4 +5,4 @@ import hashlib
 def generate_hash(url_string):
     m = hashlib.sha256()
     m.update(url_string.encode('utf-8'))
-    return m.hexdigest()[:14]
+    return unicode(m.hexdigest()[:14])
