+++ bookie/tests/factory.py
@@ -25,12 +25,12 @@ def random_string(length=None):
     """
     chars = string.ascii_uppercase + string.digits
     str_length = length if length is not None else random_int()
-    return u''.join(random.choice(chars) for x in range(str_length))
+    return unicode(u''.join(random.choice(chars) for x in range(str_length)))
 
 
 def random_url():
     """Generate a random url that is totally bogus."""
-    url = "http://{0}.com".format(random_string())
+    url = u"http://{0}.com".format(random_string())
     return url
 
 
@@ -43,11 +43,11 @@ def make_applog(message=None, status=None):
         message = random_string(100)
 
     alog = AppLog(**{
-      'user': random_string(10),
-      'component': random_string(10),
-      'status': status,
-      'message': message,
-      'payload': '',
+        'user': random_string(10),
+        'component': random_string(10),
+        'status': status,
+        'message': message,
+        'payload': u'',
     })
     return alog
 
@@ -62,7 +62,7 @@ def make_tag(name=None):
 def make_bookmark(user=None):
     """Generate a fake bookmark for testing use."""
     bmark = Bmark(random_url(),
-                  username="admin",
+                  username=u"admin",
                   desc=random_string(),
                   ext=random_string(),
                   tags=u"bookmarks")
