+++ requirements.txt
@@ -21,11 +21,11 @@ Jinja2==2.5.5
 kombu==2.5.12
 logilab-astng==0.21.1
 logilab-common==0.55.0
-lxml==3.2.4 
+lxml==3.2.4
 Mako==0.4.1
 MarkupSafe==0.12
 mock==0.7.0
-nose==1.1.2
+nose2 <= 0.5
 nose-selecttests
 Paste==1.7.5.1
 PasteDeploy==1.5.0
@@ -55,7 +55,7 @@ six==1.3.0
 Sphinx==1.0.7
 SQLAlchemy==0.8.2
 Tempita==0.5dev
-transaction==1.1.1
+transaction==1.4.1
 translationstring==0.4
 Unidecode==0.04.7
 unittest2==0.5.1
