+++ test.ini
@@ -48,7 +48,7 @@ port = 6543
 # Begin logging configuration
 
 [loggers]
-keys = root, bookie, sqlalchemy
+keys = root, bookie, celery, breadability
 
 [handlers]
 keys = console
@@ -57,16 +57,26 @@ keys = console
 keys = generic
 
 [logger_root]
-level = DEBUG
+level = INFO
 handlers = console
 
 [logger_bookie]
-level = DEBUG
+level = WARN
 handlers =
 qualname = bookie
 
+[logger_breadability]
+level = WARN
+handlers = 
+qualname = breadability
+
+[logger_celery]
+level = WARN
+handlers =
+qualname = bookie.bcelery.tasks
+
 [logger_sqlalchemy]
-level = DEBUG
+level = WARN
 handlers = console
 qualname = sqlalchemy.engine
 # "level = INFO" logs SQL queries.
@@ -76,7 +86,7 @@ qualname = sqlalchemy.engine
 [handler_console]
 class = StreamHandler
 args = (sys.stderr,)
-level = DEBUG
+level = INFO
 formatter = generic
 
 [formatter_generic]
